refactor: replace Lottie with GIF files
