--- src/components/LottieAnimation/files/index.ts
@@ -1,2 +1,0 @@
-export { default as spinner } from './spinner.json';
-export { default as greenPulse } from './green-pulse.json';
