+++ src/components/Spinner/index.tsx
@@ -1,7 +1,8 @@
 /** @jsxImportSource @emotion/react */
 import React from 'react';
+import { useTranslation } from 'translation';
 
-import { Spinner as SpinnerAnimation } from '../LottieAnimation';
+import spinnerAnimation from './spinnerAnimation.gif';
 import { useStyles } from './styles';
 import TEST_IDS from './testIds';
 
@@ -12,10 +13,11 @@ interface SpinnerProps {
 
 export const Spinner: React.FC<SpinnerProps> = ({ variant = 'large', className }) => {
   const styles = useStyles({ variant });
+  const { t } = useTranslation();
 
   return (
     <div css={styles.container} className={className} data-testid={TEST_IDS.spinner}>
-      <SpinnerAnimation css={styles.spinner} />
+      <img css={styles.spinner} src={spinnerAnimation} alt={t('spinner.altText')} />
     </div>
   );
 };
